Fix tests

+++ data/beaconrestapi/src/test/java/tech/pegasys/teku/beaconrestapi/handlers/v1/validator/GetInclusionListCommitteeDutiesTest.java
@@ -94,7 +94,7 @@ public class GetInclusionListCommitteeDutiesTest extends AbstractMigratedBeaconH
   private InclusionListDuty getInclusionListCommitteeDuty(
       final int validatorIndex, final UInt64 slot) {
     return new InclusionListDuty(
-        slot, validatorIndex, BLSTestUtil.randomPublicKey(1), inclusionListCommitteeRoot);
+        slot, UInt64.valueOf(validatorIndex), BLSTestUtil.randomPublicKey(1), inclusionListCommitteeRoot);
   }
 
   @Test
